+++ tests_metricflow/dataflow_plan_to_svg.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 import os
 
 from _pytest.fixtures import FixtureRequest
-from metricflow_semantics.dag.dag_visualization import DagGraphT, render_via_graphviz
 from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfiguration
 from metricflow_semantics.test_helpers.snapshot_helpers import snapshot_path_prefix
+from metricflow_semantics.toolkit.mf_graph.formatting.dag_visualization import DagGraphT, render_via_graphviz
 
 
 def display_graph_if_requested(
